Only compile/eval unique beta node expressions
I discovered that the truth maintenance can fail to produce predicted working memory states along certain control paths.  In this PR I have a test that exposes this issue in as small of a reproduction as I could devise right now.

This issue shows up here because the the use of an `:or` disjunction in the example test rule.  The nested conditions involved in the disjunction both have a common child ProductionNode.  When ProductionNodes have their activations added and retracted into Clara memory, the production Activation includes the entire ProductionNode under the `:node` key.  This means for one Activation to be considered equal to another Activation in memory the representation of the ProductionNode must be value-based equals (i.e. via `=`).

The problem is that ProductionNodes have a key, `:rhs`, that holds a compiled Clojure function object.  In general, it is likely a good idea that the Clara compiler should avoid compiling expressions that are equal more than one time.  For one, this has a potential benefit for both time and memory consumption.  More importantly, this is necessary behavior when the compiled Clojure function is associated to a Rete node that is used in its equality comparisons later (i.e. ProductionNodes).

What happens in the test scenario is that two nested, "sibling" conditions of the disjunction condition both have a reference to a ProductionNode that is actually the same `:id` and same rule LHS and RHS.  However, the function `clara.rules.compiler/compile-beta-tree` does not avoid compiling equalivalent expressions multiple times.  This PR fixes this specifically.

I have changed `c.r.c/compile-beta-tree` to maintain local state for "memoizing" the expression compilation (done via `eval`) calls when recursively compiling beta nodes.  This does not only fix the issue with ProductionNodes.  I have covered all compiled expressions in this process.  I have only found an issue in ProductionNodes though because of the way that an Activation refers to them and is keyed off of in Clara memory.
I do think it is a good idea to avoid any extra compilation and a good idea to ensure all nodes are completely equal, via `=`, when they represent the same node `:id` (e.g. for AccumulateWithJoinFilterNode, ProductionNode, etc).

The code changes to `c.r.c/compile-beta-tree` effected the way the forms are nested and therefore made the diff harder to see.  I think this restructure makes the function public signature clearer though and adds this memoized beta node expression compilation.  In case it is hard to see, the parts I changed were:
* Nesting the inner function recursion target.
* The inner function for memoizing compiling beta node expressions.
* All former direct uses of `eval` have been put behind this memoized variant.
* The public signature of the function no longer has `is-root` or `parent-bindings` passed in because I don't think that any consuming functions would need to be aware of those details anyways.

I will note that we could probably take this a step further (later) to have `c.r.c/compile-beta-tree` just reuse entire beta nodes if it has already compiled before along a different branch of the network.  This would be a step beyond just sharing the compiled/`eval`ed expressions beta nodes reference.  I didn't do that here just because it was potentially more disruptive.
Also, there this same sort of work may be beneficial when compiling the alpha nodes in the network.